requirements.txt
matplotlib==3.4.3
numpy==1.20.3
Pillow==9.0.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.62.3
~pencv_python==4.5.4.58
